Add build fixes for NowPlaying and OnScreenActions
